feature: disable cache config option
Cache creates problems - if manipulating tasks data directly in the database, changes are not reflected in views. I also had many problems disabling a task in the interface but the task still being run by the scheduler (it even appeared in schedule:list). So I implemented the possibility to disable cache from the config.